implement cached_property and fix circular import
