+++ posthog/models/filters/mixins/base.py
@@ -1,0 +1,5 @@
+from typing import Dict
+
+
+class BaseParamMixin:
+    _data: Dict
