+++ posthog/models/filters/path_filter.py
@@ -1,6 +1,7 @@
 from posthog.models.filters.base_filter import BaseFilter
-from posthog.models.filters.mixins.common import DateMixin, IntervalMixin, PropertyMixin
+from posthog.models.filters.mixins.common import DateMixin, IntervalMixin
 from posthog.models.filters.mixins.paths import ComparatorMixin, PropTypeMixin, StartPointMixin, TargetEventMixin
+from posthog.models.filters.mixins.property import PropertyMixin
 
 
 class PathFilter(
